Resolve header field metadata through the trace's own family
ensure_self_describing and the resolve_* fallbacks consulted the GB
legacy lookup regardless of the header's family, so a NES header's
u16 fields (line, dot) got typed u8 — and the column writer silently
dropped every mismatched set_u16, producing chunks with unequal column
lengths. The first NES end-to-end capture caught it. Resolution now
goes through family_def()'s catalogue (identical behaviour for GB,
which is also the legacy fallback), Family grows the field_group
helper this needs, and the NES cycles field widens to u16 while here —
OAM DMA freezes the CPU for 513+ cycles inside one instruction.

+++ crates/gbtrace/src/family/nes/mod.rs
@@ -25,7 +25,9 @@ pub static CPU: SubsystemDef = SubsystemDef {
             field!("rdy", bool),
         ]),
         (Layer::Timing, &[
-            field!("cycles", u8),
+            // u16: OAM DMA freezes the CPU for 513+ cycles inside one
+            // instruction, overflowing a u8 delta.
+            field!("cycles", u16),
         ]),
     ],
 };
